+++ CHANGELOG.md
@@ -12,6 +12,7 @@
 - Fixed highlighting of string literals in deprecation pragmas ([#112](https://github.com/JustusAdam/language-haskell/issues/112))
 - Specialized highlighting for `signature`, to permit it as an identifier ([#108](https://github.com/JustusAdam/language-haskell/issues/108))
 - Fixed comments occurring in place of export lists ([#104](https://github.com/JustusAdam/language-haskell/issues/104))
+- Pragmas are now case insensitive ([#81](https://github.com/JustusAdam/language-haskell/issues/81))
 
 ## 2.7.0 - 29.12.2019